# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: API Integration
# during dev, should point to your local machine
# host: localhost:5040
host: 52.39.212.226:5054
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/visitSurvey:
    # binds a127 app logic to a route
    x-swagger-router-controller: survey_ctrl
    post:
      tags: ['Survey Module']
      description: Visit Survey
      summary:  Visit Survey
      # used as the method name of the controller
      operationId: visitSurvey
      parameters:
        - name: body
          in: body
          description: Visit Survey
          schema:
           $ref: '#/definitions/visitSurvey'
          required: true       
        # - name: Authorization
        #   in: header
        #   description: Authorization
        #   required: true
        #   type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /api/submitSurvey:
    # binds a127 app logic to a route
    x-swagger-router-controller: survey_ctrl
    post:
      tags: ['Survey Module']
      description: submit Survey
      summary:  submit Survey
      # used as the method name of the controller
      operationId: submitSurvey
      parameters:
        - name: body
          in: body
          description: submit Survey
          schema:
           $ref: '#/definitions/submitSurvey'
          required: true       
        # - name: Authorization
        #   in: header
        #   description: Authorization
        #   required: true
        #   type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /api/addSurvey:
    # binds a127 app logic to a route
    x-swagger-router-controller: survey_ctrl
    post:
      tags: ['Survey Module']
      description: add Survey
      summary:  add Survey
      # used as the method name of the controller
      operationId: addSurvey
      parameters:
        - name: body
          in: body
          description:  add Survey
          schema:
           $ref: '#/definitions/addSurvey'
          required: true       
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string               
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getSurveySetByAdmin:
    # binds a127 app logic to a route
    x-swagger-router-controller: survey_ctrl
    get:
      tags: ['Survey Module']
      description: get Survey
      summary:  get Survey
      # used as the method name of the controller
      operationId: getSurveySetByAdmin      
      # parameters:
        # - name: Authorization
        #   in: header
        #   description: Authorization
        #   required: true
        #   type: string                  
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
...
  /api/getSurveyById:
    # binds a127 app logic to a route
    x-swagger-router-controller: survey_ctrl
    post:
      tags: ['Survey Module']
      description: get survay by VisitId 
      summary: get survay by VisitId 
      # used as the method name of the controller
      operationId: getSurveyByVisitId
      parameters:
         - name: body
           in: body
           description: get Survey By Id
           schema:
            $ref: '#/definitions/getSurveyById'
           required: true   
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
...
definitions:                         
...
  visitSurvey:
    type: object
    properties:
      rating:
        type: number
      feedback:
       type: array
       items:
        type: object
        properties:
          question:
           type: string
          answer:
           type: string  
  addSurvey:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      question:
        type: array
        items:
          type: object
          properties:
            description:
              type: string
            type:
              type: string
      level:
        type: array
        items:
          type: string
  # getNurseTests:
  #   type: object
  #   properties:
  #    visitId:
  #       type: string
...
  submitSurvey:
    type: object
    properties:
       _id:
        type: string 
  getSurveyById:
    type: object
    properties:
       id:
        type: string 
  appResponse:
    # required:
    #   - code
    properties:
      code:
        type: number
      status:
        type: string        
      message:
        type: string        
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  # by sunny on 15-april-2017
...